+++ src/components/projects/ProjectList.js
@@ -1,9 +1,27 @@
 import React from 'react'
 import ProjectCard from './ProjectCard'
 
-const ProjectList = ({ projects, showTags = true, showStars = true, showDelta = true, showDescription = true, showURL = false, showMetrics = true, deltaFilter, viewOptions }) => {
-  return (
-    <div className="project-grid">
+class ProjectList extends React.Component {
+  shouldComponentUpdate () {
+    // Never re-render the project list
+    // to avoid rendering after successful login, and fetch events about heroes, links, reviews and requests
+    return false
+  }
+  render () {
+    const {
+      projects,
+      showTags = true,
+      showStars = true,
+      showDelta = true,
+      showDescription = true,
+      showURL = false,
+      showMetrics = true,
+      deltaFilter,
+      viewOptions
+    } = this.props
+    console.info('Render the list', projects[0])
+    return (
+      <div className="project-grid">
       {projects.map((project, index) =>
         <div key={project.slug}>
           <ProjectCard
@@ -20,7 +38,9 @@ const ProjectList = ({ projects, showTags = true, showStars = true, showDelta = 
           />
         </div>)
       }
-    </div>
-  )
+      </div>
+    )
+  }
 }
+
 export default ProjectList